+++ thermion_dart/native/src/c_api/TRenderManager.cpp
@@ -50,6 +50,11 @@ EMSCRIPTEN_KEEPALIVE void RenderManager_requestRender(TRenderManager *tRenderer)
     renderManager->requestRender();
 }
 
+EMSCRIPTEN_KEEPALIVE void RenderManager_setTargetFps(TRenderManager *tRenderer, int fps) {
+    auto *renderManager = reinterpret_cast<RenderManager *>(tRenderer);
+    renderManager->setTargetFps(fps);
+}
+
 EMSCRIPTEN_KEEPALIVE void RenderManager_setPaused(TRenderManager *tRenderer, bool paused) {
     auto *renderManager = reinterpret_cast<RenderManager *>(tRenderer);
     renderManager->setPaused(paused);